Add ability to tag a demo with domain(s) Issue #44
Allow an authenticated user to associate a new or existing demo with a
domain (DoD, Intel, FinSiv, etc) and display the domains as facets.
Requires a new range path index.

Refactored features directive.  Renamed to be items to be more generic
and added a label to its scope.  This should be  more flexible if
another features-like attribute is required to be tagged on a demo.
Just create a new service with the seed values.