#100 & #88 Add Points of Contact
Add a new persons array to a demo, displayed as Points of Contact .
Allow three roles: Technical Contact, Business Owner, External Contact.
 This supports n number of persons with n number of roles.

Migrate current maintainer and email to that of a Technical Contact.
Provided XQuery module maintainer-to-tech-contact.xqy will preform this
migration.

Remove maintainer element range index and replace with that for a
personName.  Update search options for new facet.

Update the emails sent to the appropriate person: Technical Contact for
bug and Business Owner for comment.